feat: add cloud infrastructure network request events UI
- Update cloud infra template to use bandwidth_bytes (10M-100M range) instead of bandwidth_MB
- Create modular event form components:
  - CloudInfraEventsForm for bandwidth_bytes and runtime fields
  - AiAgentsEventsForm for AI agent-specific fields
  - GenericEventsForm for simple event sending
  - Refactor EventsCard to use composition pattern
- Update EventsCard to get instance from Zustand store instead of props
- Update all tests to use new bandwidth_bytes property and ranges
- Maintain backward compatibility with existing AI agents functionality
